开始单元测试

+++ src/SQLiteLib/Shared/DataLib.Shared/Table/Impl/DataRow.cs
@@ -17,7 +17,7 @@ namespace DataLib.Table.Impl
         /// <summary>
         /// Row Index
         /// </summary>
-        public ulong RowIndex { get; set; }
+        public int RowIndex { get; set; }
 
         /// <summary>
         /// 数据结果集合
